Added ignoreCase option
Added a flag to make searches case-insensitive.

Could use a minor aesthetic fix: the background hint of the first matched entry will still be its original case. If possible, the hint should match the case entered, but the item (as would be entered by pressing the right arrow) should remain as it always was.